\name{cnps}
\alias{cnps}
\docType{data}
\title{Segment-by-cell matrix of copy number states from NCI-N87 cell line.}
\description{Matrix of segments (rows) X 200 cells (columns) with entries denoting inferred copy numbers.}
\usage{data(cnps)}

\source{Data obtained from Ji lab at Stanford.}

\keyword{datasets}
