\name{eps}
\alias{eps}
\docType{data}
\title{Segment-by-cell matrix of expression from NCI-N87 cell line.}
\description{Matrix of segments (rows) x 200 cells (columns) with entries denoting average expression values.}
\usage{data(eps)}

\source{Data obtained from Ji lab at Stanford.}

\keyword{datasets}
